+++ graylog2-web-interface/src/components/users/UserDetails/ProfileSection.jsx
@@ -3,8 +3,8 @@ import * as React from 'react';
 
 import User from 'logic/users/User';
 import SectionComponent from 'components/common/Section/SectionComponent';
+import ReadOnlyFormGroup from 'components/common/ReadOnlyFormGroup';
 
-import ReadOnlyFormGroup from '../form/ReadOnlyFormGroup';
 import LoggedInIcon from '../LoggedInIcon';
 
 type Props = {
