+++ graylog2-web-interface/src/components/users/UserDetails/SettingsSection.jsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import User from 'logic/users/User';
 import SectionComponent from 'components/common/Section/SectionComponent';
 
-import ReadOnlyFormGroup from '../form/ReadOnlyFormGroup';
+import ReadOnlyFormGroup from '../../common/ReadOnlyFormGroup';
 
 type Props = {
   user: User,
