+++ graylog2-web-interface/src/components/users/UserEdit/ProfileSection.jsx
@@ -5,8 +5,8 @@ import { Formik, Form } from 'formik';
 import { Button, Col, Row } from 'components/graylog';
 import User from 'logic/users/User';
 import SectionComponent from 'components/common/Section/SectionComponent';
+import ReadOnlyFormGroup from 'components/common/ReadOnlyFormGroup';
 
-import ReadOnlyFormGroup from '../form/ReadOnlyFormGroup';
 import FullNameFormGroup from '../UserCreate/FullNameFormGroup';
 import EmailFormGroup from '../UserCreate/EmailFormGroup';
